// frontend/src/types/random.d.ts
export interface IMealItem {
    name: string;
    ingredients: string[];
    specialInstructions: string[];
    _id?: string;
}
export type MealTimeStatus = "pending" | "preparing" | "delivered";
export type MealTime = "morning" | "evening" | "night";
export interface IMealItem {
    name: string;
    ingredients: string[];
    specialInstructions: string[];
}
export interface IMeal {
    _id: string;
    date: string;
    meals: {
        morning: {
            items: IMealItem[];
            status: string;
            _id: string;
        };
        evening: {
            items: IMealItem[];
            status: string;
            _id: string;
        };
        night: {
            items: IMealItem[];
            status: string;
            _id: string;
        };
    };
    patientId: {
        _id: string;
        name: string;
        roomNumber: string;
        bedNumber: string;
    };
}
export interface IMeals {
    meal: IMeal;
    className?: string;
}